Bug fixed on nb_queue_node_get_prev()

+++ include/nb/memory_bot/membank.h
@@ -12,6 +12,6 @@ void nb_membank_data_free(nb_membank_t *membank, void *obj);
 void nb_membank_finish(nb_membank_t *membank);
 void nb_membank_clear(nb_membank_t *membank);
 void nb_membank_merge(nb_membank_t *membank, nb_membank_t *extension);
-void nb_membank_set_N_x_block(nb_membank_t *membank, int N_x_block);
+void nb_membank_set_N_x_block(nb_membank_t *membank, uint16_t N_x_block);
 
 #endif
